+++ packages/falso/src/index.ts
@@ -193,3 +193,11 @@ export { randAirport } from './lib/airport';
 export { randAirportName } from './lib/airport-name';
 export { randAirportCode } from './lib/airport-code';
 export { randCreditCard } from './lib/credit-card';
+export {
+  incrementalNumber,
+  IncrementalNumberOptions,
+} from './lib/factories/incremental-number';
+export {
+  incrementalDate,
+  IncrementalDateOptions,
+} from './lib/factories/incremental-date';