+++ packages/falso/src/lib/between-date.ts
@@ -3,8 +3,8 @@ import { randNumber } from './number';
 import { dateIsUnique } from './core/unique-validators';
 
 interface BetweenOptions extends FakeOptions {
-  from: Date;
-  to: Date;
+  from: Date | string;
+  to: Date | string;
 }
 
 /**
@@ -28,8 +28,8 @@ interface BetweenOptions extends FakeOptions {
 export function randBetweenDate<Options extends BetweenOptions = never>(
   options: Options
 ) {
-  const from = options.from.getTime();
-  const to = options.to.getTime();
+  const from = new Date(options.from).getTime();
+  const to = new Date(options.to).getTime();
 
   if (from >= to) {
     throw new Error('From must be before to');