+++ packages/falso/src/tests/between-date.spec.ts
@@ -11,6 +11,16 @@ describe('randBetweenDate', () => {
     expect(result.getTime()).toBeLessThan(to.getTime());
   });
 
+  it('should use strings for dates', () => {
+    const from = new Date().toISOString();
+    const to = new Date('12/02/2022').toISOString();
+
+    const result = randBetweenDate({ from, to });
+
+    expect(result.getTime()).toBeGreaterThan(new Date(from).getTime());
+    expect(result.getTime()).toBeLessThan(new Date(to).getTime());
+  });
+
   describe('from date is after to date', () => {
     let from: Date;
     let to: Date;